requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-frontmatter==1.0.0
markdown==3.5.1
Jinja2==3.1.3